+++ optimization/forwardModelKinetics.py
@@ -1,9 +1,8 @@
 import math
 import numpy as np
-import pandas as pd
 import torch
 import math as math
-from jax import numpy as jnp
+
 
 
 def forwardModelKinetics(
